+++ src/components/Inputs/RichTextEditor/MenuBar/Alignment.tsx
@@ -1,19 +1,20 @@
+import { FC } from 'react';
 import {
   format_align_center,
   format_align_left,
   format_align_right,
 } from '@equinor/eds-icons';
 import { EditorPanel, RichTextEditorFeatures } from '../RichTextEditor.types';
-import { Section } from './MenuBar.styles';
+import { MenuSection } from './MenuBar.styles';
 import MenuButton from './MenuButton';
 
-const Alignment = ({ editor, features }: EditorPanel) => {
+const TextAlignment: FC<EditorPanel> = ({ editor, features }) => {
   const alignLeft = () => editor.chain().focus().setTextAlign('left').run();
   const alignCenter = () => editor.chain().focus().setTextAlign('center').run();
   const alignRight = () => editor.chain().focus().setTextAlign('right').run();
   if (!features.includes(RichTextEditorFeatures.ALIGNMENT)) return;
   return (
-    <Section>
+    <MenuSection>
       <MenuButton
         active={editor.isActive({ textAlign: 'left' })}
         icon={format_align_left}
@@ -29,8 +30,8 @@ const Alignment = ({ editor, features }: EditorPanel) => {
         icon={format_align_right}
         onClick={alignRight}
       />
-    </Section>
+    </MenuSection>
   );
 };
 
-export { Alignment };
+export { TextAlignment };
