+++ src/components/Inputs/RichTextEditor/MenuBar/Code.tsx
@@ -1,9 +1,10 @@
+import { FC } from 'react';
 import { code } from '@equinor/eds-icons';
 
 import { EditorPanel, RichTextEditorFeatures } from '../RichTextEditor.types';
 import MenuButton from './MenuButton';
 
-const Code = ({ editor, features }: EditorPanel) => {
+const TextCode: FC<EditorPanel> = ({ editor, features }) => {
   const toggleCode = () => editor.chain().focus().toggleCodeBlock().run();
   if (!features.includes(RichTextEditorFeatures.CODE)) return;
   return (
@@ -15,4 +16,4 @@ const Code = ({ editor, features }: EditorPanel) => {
   );
 };
 
-export { Code };
+export { TextCode };
