+++ src/components/Inputs/RichTextEditor/MenuBar/History.tsx
@@ -1,15 +1,16 @@
+import { FC } from 'react';
 import { redo, undo } from '@equinor/eds-icons';
 
 import { EditorPanel, RichTextEditorFeatures } from '../RichTextEditor.types';
-import { Section } from './MenuBar.styles';
+import { MenuSection } from './MenuBar.styles';
 import MenuButton from './MenuButton';
 
-const History = ({ editor, features }: EditorPanel) => {
+const TextHistory: FC<EditorPanel> = ({ editor, features }) => {
   const onUndo = () => editor.chain().focus().undo().run();
   const onRedo = () => editor.chain().focus().redo().run();
   if (!features.includes(RichTextEditorFeatures.HISTORY)) return;
   return (
-    <Section>
+    <MenuSection>
       <MenuButton
         icon={undo}
         onClick={onUndo}
@@ -20,8 +21,8 @@ const History = ({ editor, features }: EditorPanel) => {
         onClick={onRedo}
         disabled={!editor.can().redo()}
       />
-    </Section>
+    </MenuSection>
   );
 };
 
-export { History };
+export { TextHistory };
