+++ src/components/Inputs/RichTextEditor/MenuBar/Lists.tsx
@@ -1,17 +1,18 @@
+import { FC } from 'react';
 import { format_list_bulleted, format_list_numbered } from '@equinor/eds-icons';
 
 import { EditorPanel, RichTextEditorFeatures } from '../RichTextEditor.types';
-import { Section } from './MenuBar.styles';
+import { MenuSection } from './MenuBar.styles';
 import MenuButton from './MenuButton';
 
-const Lists = ({ editor, features }: EditorPanel) => {
+const TextLists: FC<EditorPanel> = ({ editor, features }) => {
   const toggleBulletList = () =>
     editor.chain().focus().toggleBulletList().run();
   const toggleOrderedList = () =>
     editor.chain().focus().toggleOrderedList().run();
   if (!features.includes(RichTextEditorFeatures.LISTS)) return;
   return (
-    <Section>
+    <MenuSection>
       <MenuButton
         active={editor.isActive('bulletList')}
         icon={format_list_bulleted}
@@ -22,8 +23,8 @@ const Lists = ({ editor, features }: EditorPanel) => {
         icon={format_list_numbered}
         onClick={toggleOrderedList}
       />
-    </Section>
+    </MenuSection>
   );
 };
 
-export { Lists };
+export { TextLists };
